Criação do post para oocupação Solicitação de compra

+++ ApiSGCOlimpiada/Models/OcupacaoSolicitacaoCompra.cs
@@ -7,8 +7,8 @@ namespace ApiSGCOlimpiada.Models
 {
     public class OcupacaoSolicitacaoCompra
     {
-        public int OcupacaoId { get; set; }
-        public int SolicitacaoId { get; set; }
+        public long OcupacaoId { get; set; }
+        public long SolicitacaoId { get; set; }
         public Ocupacao Ocupacao { get; set; }
         public SolicitacaoCompra SolicitacaoCompra { get; set; }
     }
